CSVParserSourceFunction should be checkpointed
Former-commit-id: 8b1504f09e666f31bb1eda67f9d9ea95397008ad

+++ examples/src/main/java/com/spbsu/flamestream/example/benchmark/LentaBenchStand.java
@@ -96,7 +96,7 @@ public class LentaBenchStand {
     BenchValidator<Prediction> validator =
             ((Class<? extends BenchValidator>) Class.forName(validatorClass)).newInstance();
     final AwaitCountConsumer awaitConsumer = new AwaitCountConsumer(validator.expectedOutputSize());
-    final Map<Integer, LatencyMeasurer> latencies = Collections.synchronizedMap(new LinkedHashMap<>());
+    final Map<Long, LatencyMeasurer> latencies = Collections.synchronizedMap(new LinkedHashMap<>());
     try (
             AutoCloseable ignored = benchStandBuilder.producer(
                     TextDocument.class,
